support for multiple examples
