+++ package.json
@@ -53,6 +53,7 @@
     "material-ui": "^1.0.0-beta.36",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-router-dom": "^4.2.2",
     "strip-loader": "^0.1.2",
     "validator": "^9.4.1",
     "webpack": "^4.1.1",
